+++ server/nmtwizard/config.py
@@ -16,8 +16,7 @@ def add_log_handler(fh):
 def merge_config(a, b, name):
     if isinstance(a, dict):
         for k in six.iterkeys(b):
-            if k not in a or not(
-               (isinstance(a[k], dict) and isinstance(b[k], dict))):
+            if k not in a or type(a[k]) != type(b[k]):
                 a[k] = b[k]
             elif isinstance(a[k], dict):
                 merge_config(a[k], b[k], name)